CH-152 user attributes generic API
